Filtrado por año, y año temporal por sesion. agregar a parametro ano_activo y agregar a temp : temp_iduser = 1.

+++ protected/controllers/TempController.php
@@ -1,6 +1,6 @@
 <?php
 
-class CursoController extends Controller
+class TempController extends Controller
 {
 	/**
 	 * @var string the default layout for the views. Defaults to '//layouts/column2', meaning
@@ -62,16 +62,16 @@ class CursoController extends Controller
 	 */
 	public function actionCreate()
 	{
-		$model=new Curso;
+		$model=new Temp;
 
 		// Uncomment the following line if AJAX validation is needed
 		// $this->performAjaxValidation($model);
 
-		if(isset($_POST['Curso']))
-		{
-			$model->attributes=$_POST['Curso'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->cur_id));
+		if (isset($_POST['Temp'])) {
+			$model->attributes=$_POST['Temp'];
+			if ($model->save()) {
+				$this->redirect(array('view','id'=>$model->temp_id));
+			}
 		}
 
 		$this->render('create',array(
@@ -91,11 +91,11 @@ class CursoController extends Controller
 		// Uncomment the following line if AJAX validation is needed
 		// $this->performAjaxValidation($model);
 
-		if(isset($_POST['Curso']))
-		{
-			$model->attributes=$_POST['Curso'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->cur_id));
+		if (isset($_POST['Temp'])) {
+			$model->attributes=$_POST['Temp'];
+			if ($model->save()) {
+				$this->redirect(array('view','id'=>$model->temp_id));
+			}
 		}
 
 		$this->render('update',array(
@@ -110,11 +110,17 @@ class CursoController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
+		if (Yii::app()->request->isPostRequest) {
+			// we only allow deletion via POST request
+			$this->loadModel($id)->delete();
 
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+			if (!isset($_GET['ajax'])) {
+				$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			}
+		} else {
+			throw new CHttpException(400,'Invalid request. Please do not repeat this request again.');
+		}
 	}
 
 	/**
@@ -122,26 +128,7 @@ class CursoController extends Controller
 	 */
 	public function actionIndex()
 	{
-                
-                $par = Parametro::model()->findByAttributes(array('par_item'=>'ano_activo'));
-                $temp = Temp::model()->findByAttributes(
-                     array('temp_iduser'=>Yii::app()->user->id)
-                 );      
-            
-                if ( $temp->temp_ano != 0 ){
-                      $ano = array($temp->temp_ano);
-                } else {
-                    $ano = array($par->par_descripcion);
-                }
-                
-                $criteria = new CDbCriteria();
-                $criteria->addInCondition('cur_ano', $ano, 'or');
-                
-                
-		$dataProvider=new CActiveDataProvider('Curso', array(
-                'criteria' => $criteria
-                ));
-                
+		$dataProvider=new CActiveDataProvider('Temp');
 		$this->render('index',array(
 			'dataProvider'=>$dataProvider,
 		));
@@ -152,10 +139,11 @@ class CursoController extends Controller
 	 */
 	public function actionAdmin()
 	{
-		$model=new Curso('search');
+		$model=new Temp('search');
 		$model->unsetAttributes();  // clear any default values
-		if(isset($_GET['Curso']))
-			$model->attributes=$_GET['Curso'];
+		if (isset($_GET['Temp'])) {
+			$model->attributes=$_GET['Temp'];
+		}
 
 		$this->render('admin',array(
 			'model'=>$model,
@@ -166,27 +154,32 @@ class CursoController extends Controller
 	 * Returns the data model based on the primary key given in the GET variable.
 	 * If the data model is not found, an HTTP exception will be raised.
 	 * @param integer $id the ID of the model to be loaded
-	 * @return Curso the loaded model
+	 * @return Temp the loaded model
 	 * @throws CHttpException
 	 */
 	public function loadModel($id)
 	{
-		$model=Curso::model()->findByPk($id);
-		if($model===null)
+		$model=Temp::model()->findByPk($id);
+		if ($model===null) {
 			throw new CHttpException(404,'The requested page does not exist.');
+		}
 		return $model;
 	}
 
 	/**
 	 * Performs the AJAX validation.
-	 * @param Curso $model the model to be validated
+	 * @param Temp $model the model to be validated
 	 */
 	protected function performAjaxValidation($model)
 	{
-		if(isset($_POST['ajax']) && $_POST['ajax']==='curso-form')
-		{
+		if (isset($_POST['ajax']) && $_POST['ajax']==='temp-form') {
 			echo CActiveForm::validate($model);
 			Yii::app()->end();
 		}
 	}
+        
+        public function actionAjax_ano(){
+            $ano = $_POST['Temp'];
+            
+        }
 }
